chore(framework): move package internal to pkg/utils
