+++ pkg/utils/kubeconfig.go
@@ -1,4 +1,4 @@
-package internal
+package utils
 
 import (
 	"github.com/spf13/cobra"
