+++ pkg/utils/vars.go
@@ -1,4 +1,4 @@
-package internal
+package utils
 
 import "time"
